Task Alex, componente Autocomplete

+++ Shared/Models/Produto.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -11,7 +12,7 @@ namespace BlazorEcomm.Shared.Models
         public int Id { get; set; }
         public string Nome { get; set; } = "";
         public double Preco { get; set; }
-        public double Quantidade { get; set; }
+        public int Quantidade { get; set; }
         public string Imagem { get; set; } = "";
     }
 }
